When reading sensors, treat ENOENT the same as ENXIO.
This allows the user to specify sensors that don't necessarily exist.

As an example, the k10temp driver reports the Tccd1 temp sensor on only
one device, but there's no way to know which one it is. With this commit,
the user can specify the sensor to be reported from both devices.