Change `dev_path` logic to be more robust.
Before this commit, sensord looked for the hwmon directory under
the given `dev_path`. This was problematic for some cases like `nvme` nodes
where the node incorporated a number that depended on when the device was
probed. For example the path could be:

    /sys/devices/pci0000:00/0000:00:01.2/0000:04:00.0/nvme/nvme0

... on one boot and:

    /sys/devices/pci0000:00/0000:00:01.2/0000:04:00.0/nvme/nvme1

... on the next.

With this commit, sensord now instead resolves every hwmon node in
`/sys/class/hwmon` first, and then matches them up to the configured
`dev_name` and `dev_path` entries. For `dev_name` entries it picks the first
hwmon node that has a matching `name.` For `dev_path` entries it picks
the first hwmon node whose canonicalized path lies under that path.

So in the above example the user can now specify:

    dev_path = "/sys/devices/pci0000:00/0000:00:01.2/0000:04:00.0"

... to get the same effect.